requirements.txt
flask
flask-cors
groq
sentence-transformers
faiss-cpu
numpy
